# Conflicts:
#   frontend/moa/src/api/quizApi.js
#   frontend/moa/src/components/Quiz/QuestionArea.jsx
#   frontend/moa/src/pages/Quiz/IncorrectNote.jsx
#   frontend/moa/src/pages/Quiz/SentenceQuiz.jsx

+++ backend/moa/src/main/java/com/ssafy/moa/api/service/BuddyService.java
@@ -9,5 +9,5 @@ public interface BuddyService {
     Long findMatchingBuddy(Long memberId);
     Integer deleteBuddy(Long memberId);
     Long findWithBuddyDate(Long memberId);
-    Integer findBuddy(Member member);
+    Long findBuddy(Member member);
 }
